Complete movie posters section
The main fragment displays move posters that fetches from the movie DB API. We can display two types of movies. First one is popular movies and the second one is top rated movies. The type is selected from the menu items.

The change mainly consists of background source code that is for fetching and displaying movies data.