Implement favorite movies module
The module allows users to mark a movie as a favorite in the details views by tapping a button(star). This is for a local movies collection that does not require an API request.

The change also contains modifying the existing sorting criteria for the main view to include an additional pivot to show their favorites collection.